Bind ssh on a different IP
We use 127.255.255.254 as a dedicated IP for autosocks. This has the
effect, that we don't use 2 ports on 127.0.0.1 per connection, but wwe
use 2 IPs and bind 1 port on each per connection.

Also, the "installation procedure" has slightly changed. Dropping the
templates into systemd's main directory seems to confuse it, probably
had some local state when I tested which made it work previously.